fix(headless): fix invalid number of results for recommendations
https://coveord.atlassian.net/browse/KIT-1701
