+++ packages/headless/src/features/product-recommendations/product-recommendations-actions.ts
@@ -14,7 +14,7 @@ import {
   nonEmptyString,
 } from '../../utils/validate-payload';
 import {ArrayValue, NumberValue, StringValue} from '@coveo/bueno';
-import {getVisitorID, historyStore} from '../../api/analytics/analytics';
+import {getVisitorID, historyStore} from '../../api/analytics/search-analytics';
 import {ProductRecommendationsRequest} from '../../api/search/product-recommendations/product-recommendations-request';
 import {
   ProductRecommendation,
