+++ packages/headless/src/features/product-recommendations/product-recommendations-analytics.actions.ts
@@ -1,15 +1,11 @@
-import {
-  ProductRecommendationAnalyticsProvider,
-  StateNeededByProductRecommendationsAnalyticsProvider,
-} from '../../api/analytics/product-recommendations-analytics';
 import {AnalyticsType, makeAnalyticsAction} from '../analytics/analytics-utils';
 
 export const logProductRecommendations = makeAnalyticsAction(
   'analytics/productrecommendations/load',
   AnalyticsType.Search,
-  (client) => client.logRecommendationInterfaceLoad(),
-  (state) =>
+  (client) => client.logRecommendationInterfaceLoad()
+  /*(state) =>
     new ProductRecommendationAnalyticsProvider(
       state as StateNeededByProductRecommendationsAnalyticsProvider
-    )
+    )*/
 );
