+++ packages/headless/src/features/result-preview/result-preview-request-builder.ts
@@ -1,4 +1,4 @@
-import {getVisitorID} from '../../api/analytics/analytics';
+import {getVisitorID} from '../../api/analytics/search-analytics';
 import {
   HtmlRequest,
   HtmlRequestOptions,
